suppression de fonctions inutiles dans resize_screen/resize.c et ajout clear_screen/
