+++ travaux_preparatoires/clear_screen/compile.sh
@@ -1,0 +1,1 @@
+gcc -Wall -Wextra -Werror *.c -ltermcap
